Fakegato dependencies and definitions
