+++ package.json
@@ -60,6 +60,7 @@
     "emotion": "^11.0.0",
     "evergreen-ui": "^5.1.2",
     "express": "^4.17.1",
+    "fakegato-history": "^0.6.1",
     "http2": "^3.3.7",
     "js-big-decimal": "^1.2.4",
     "lodash": "^4.17.20",
@@ -73,8 +74,8 @@
     "serve-static": "^1.14.1",
     "winston": "^3.3.3",
     "ws": "^7.4.2",
-    "zigbee-herdsman": "^0.13.75",
-    "zigbee-herdsman-converters": "^14.0.76"
+    "zigbee-herdsman": "^0.13.79",
+    "zigbee-herdsman-converters": "^14.0.85"
   },
   "devDependencies": {
     "@emotion/eslint-plugin": "^11.2.0",
